reorganize and fix broken tests
